[mod] #214 탈퇴 이벤트 전송 로직 fragment -> activity로 이동
