+++ app/src/main/java/org/keepgoeat/presentation/my/withdraw/WithdrawActivity.kt
@@ -13,6 +13,7 @@ import kotlinx.coroutines.flow.onEach
 import org.keepgoeat.R
 import org.keepgoeat.databinding.ActivityWithdrawBinding
 import org.keepgoeat.presentation.my.MyViewModel
+import org.keepgoeat.util.UiState
 import org.keepgoeat.util.binding.BindingActivity
 import org.keepgoeat.util.extension.showKeyboard
 import org.keepgoeat.util.setVisibility
@@ -74,7 +75,7 @@ class WithdrawActivity : BindingActivity<ActivityWithdrawBinding>(R.layout.activ
             if (binding.etOtherReason.text.isNullOrBlank() && viewModel.isOtherReasonSelected.value)
                 binding.tvOtherReasonErrorMsg.setVisibility(true)
             else
-                WithdrawDialogFragment(viewModel.getWithdrawReasons()).show(
+                WithdrawDialogFragment().show(
                     supportFragmentManager,
                     "withdrawDialog"
                 )
@@ -99,6 +100,24 @@ class WithdrawActivity : BindingActivity<ActivityWithdrawBinding>(R.layout.activ
         viewModel.isKeyboardVisible.flowWithLifecycle(lifecycle).onEach { isVisible ->
             binding.rvWithdraw.setVisibility(!isVisible)
         }.launchIn(lifecycleScope)
+        viewModel.deleteAccountUiState.flowWithLifecycle(lifecycle).onEach {
+            when (it) {
+                is UiState.Success -> sendDeleteAccountEvent()
+                else -> {}
+            }
+        }.launchIn(lifecycleScope)
+    }
+
+    private fun sendDeleteAccountEvent() {
+        val reasons: MutableMap<String, Any> = mutableMapOf()
+        viewModel.getWithdrawReasons().map {
+            if (it.key == SUBJECTIVE_ISSUE) {
+                reasons[it.key] = it.value
+            } else {
+                reasons[it.key] = getString(it.value as Int)
+            }
+        }
+        viewModel.sendDeleteAccountEvent(reasons)
     }
 
     private fun clearFocus() {
@@ -114,5 +133,9 @@ class WithdrawActivity : BindingActivity<ActivityWithdrawBinding>(R.layout.activ
     override fun onStop() {
         super.onStop()
         rootView.viewTreeObserver.removeOnGlobalLayoutListener(onGlobalListener)
+    }
+
+    companion object {
+        private const val SUBJECTIVE_ISSUE = "SUBJECTIVE_ISSUE"
     }
 }
